exception

+++ src/main/java/com/example/demo/controller/UserController.java
@@ -3,6 +3,7 @@ package com.example.demo.controller;
 import com.example.demo.domain.User;
 import com.example.demo.repository.UserRepository;
 import com.example.demo.service.UserService;
+import com.example.demo.service.exception.idInvalidException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -36,8 +37,16 @@ public class UserController {
         return ResponseEntity.status(HttpStatus.CREATED).body(user1);
     }
 
+    @ExceptionHandler (value = idInvalidException.class)
+    public ResponseEntity<String> handelIdInvalidException(idInvalidException idInvalidException) {
+        return ResponseEntity.badRequest().body(idInvalidException.getMessage());
+    }
+
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable("id") long id) {
+    public ResponseEntity<String> deleteUser(@PathVariable("id") long id) throws idInvalidException {
+        if (id > 1500) {
+            throw new idInvalidException("id khong dc lon hon 1500");
+        }
         this.userService.deleteUser(id);
         return ResponseEntity.status(HttpStatus.OK).body("delete success");
     }
